+++ src/main/java/cl/duoc/demoJPA/repository/DoctorRepository.java
@@ -7,7 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
+public interface DoctorRepository extends JpaRepository<Doctor, Long> {
     List<Doctor> findByName(String name);
     List<Doctor> findByNameAndLastName(String name, String lastName);
+
+
 }
